restructure helpers as utils
