+++ game/utils/__init__.py
@@ -1,0 +1,6 @@
+""" Contains helper functions and classes to speed up development"""
+
+from .text import Text
+from .button import Button
+
+__all__ = ["Text", "Button"]
